+++ src/main/java/ru/yandex/practicum/filmorate/storage/MpaDbStorage.java
@@ -49,7 +49,7 @@ public class MpaDbStorage implements Storage<Mpa> {
             return ps;
         }, keyHolder);
 
-        Integer id = keyHolder.getKeyAs(Integer.class);
+        Long id = keyHolder.getKeyAs(Long.class);
 
         if (id != null) {
             entity.setId(id);
